Rework country simplification (#78)
* Rework country simplification (#77)

We now use lists of words and drop terms rather than trying to do
lots of string mutations that might give us different values
that aren't in the list of terms.

* Rework to reduce iteration complexity

Single word terms can be dropped in a single pass

* Fix typo and reduce iteration further

Take the first non-dropped word rather than checking all words for
whether they need to be dropped and then taking the first from
what remains.

* Test and fix empty string and all removed cases

* Handle additional whitespace

* Simplify the logic further

We can loop over an enumerated list and break when we find a word.
Multi-part words do index-based access but it's rare enough that the
performance hit for indexing an array should be low.

* Linting

* Add more comments and optimise

Moved the term building inside the IF block because we don't need
to build the lists if we don't have any candidate words!

* Fix test failure